+++ G2 - Copy (2)/backend/models/Bin.js
@@ -1,15 +1,6 @@
 import mongoose from "mongoose";
 import AutoIncrementFactory from "mongoose-sequence";
 
-// Comment Schema
-const commentSchema = new mongoose.Schema({
-  // comment_id: { type: Number, unique: true }, // Auto-incrementing comment_id
-  user: { type: mongoose.Schema.Types.ObjectId, ref: "User", required: true }, // Reference to User
-  bin: { type: mongoose.Schema.Types.ObjectId, ref: "Bin", required: true }, // Reference to Bin
-  text: { type: String, required: true }, // Comment text
-  createdAt: { type: Date, default: Date.now }, // Timestamp
-});
-
 // Bin Schema
 const binSchema = new mongoose.Schema({
   // bin_id: { type: Number, unique: true }, // Auto-incrementing bin_id
@@ -25,6 +16,5 @@ const binSchema = new mongoose.Schema({
 });
 
 const Bin = mongoose.model("Bin", binSchema);
-const Comment = mongoose.model("Comment", commentSchema);
 
-export { Bin, Comment };
+export { Bin };
